dns test scenarios. need new arwen
